Added information to each resource
Why:
Checked that all resources are correct and up to date. Information added included addresses which will be useful for the user. Added space for a map view so that users can get a visual cue for each resource. Started to move each resource information into a hash. This will reduce the size of the view file.

How:
Information for each resource is now going to be stored in an array of hashes in the controller.