Resource cards cleaned up.
How:

Maps were positioned to be fixed to the bottom of the resource card. Height of maps was also fixed for a uniform look.
New css class resource-cards-small was created for rat sitting and rat guide links. This allows for the same style to be carried through with adaptions for the reduced content.