Adoption Application fields added
Why:
More questions required on the application form.

This PR:
Rolled back previous migrations and columns were added to the create table migration. This was to reduce the number of files with similar names and timestamps.
Columns were added and form fields created to obtain information the rescue needs from potential adopters.
A helper method has been created to list all the states of america so a dropdown could be provided for the state category