Added validations to adoption application and added a font
Why:
It is important that a valid email address is provided for a potential adopter.

This commit:
Adds a regex for a valid email and this is included in the validation in the model.
Begins to add some styling for the adoption form. The font for the header of the form and the nav bar have been customized.