Dependent fields added
Why:
Certain fields are only necessary dependent on previous questions

This commit:
Adds the questions on which other questions are dependent. It adds underscore_rails gem on which dependent_fields_rails is dependent. Removed code using foundation as the app is no longer using it.
Began formatting the application form.