Adoption application can be submitted
Why:
Applications need to be saved.

This commit:
Completes the styling of the form. Two questions now have a text area which is larger than the text fields for the other questions.
Create method has been defined. If the application is complete it can be saved and the user is redirected to the adopt page. Otherwise they are returned to their form.
Flash messages set up.