Error messages display on form
Why: It is helpful if the user is directed to the fields which are preventing their form from being submitted.

This commit:
Adds error messages to the fields which are validated.
The error messages display at the top of the form when the user is redirected back to the form.