Form labels corrected and extracted into a YAML file.
Why:
The form labels are brief descriptions but do not suffice as a question for the applicant.

This commit:
The form labels were changed to match the symbol of the form field as this is the correct form.
Using a label helper the questions which are desired were extracted into a YAML file. The YAML is called form_labels.yml
A few syntax corrections made as suggested by Hound.