Corrections following migration merge and tweeks to css for adoption application
Why:
It is important that certain fields are validated so that an empty form cannot be submitted.

This commit:
Adds validations to the model which was empty due to the rollback and new generation of the model in a previous commit to merge the two migrations.
Css tweeks to add more padding around the radio buttons and labels.
Color for fonts on the error messages matches the flash message color.
Added a "please select" for the state, this is the new default rather than Alabama.