Clean up and CSS tweaks
Why:
Adding in bitters brought default styling. Resource cards were not clear as text was too large. Overall, just wanted to customize the product more.

How:
Removed ; from the adoption_application.js. This was a recommended change however it broke the function.
Imported a new font "Playfair Display". This was used for the main body of text.
Changed line height to 1.25em. This helps with the text to fit on the resource cards.
Added extra padding-left to the resource card <p> so that it is inline with the title of the card.
Text with link to adoption application has been given the class text so that it aligns with the other text on the page.
Margin-bottom for the nav-bar has been increased to give more space between the nav-bar and headers
Used columns to position the nav menu which is more sustainable than previous method.