Removed redundancies and cleaned up routes
Why:
Small, simple and consistent is better. I want the app to be RESTful and easy to read and adapt.

This commit:
Moved CSS for the home page to it's own file.
Deleted about.html.erb as the content is included on the home page.
The view for the home page is now homes/show.html.erb
Generated a homes controller with a show action.
Renamed the routes so that the root is now homes#show.
Deleted extra routes that are not needed at this time.