Font uniformity and pictures added on some resource cards.
Why:
Some vet resource cards don't have maps due to being mobile or having multiple locations. Their cards look empty it would be nice to add something there.

This commit:
Pulls the two fonts into the variables file in base.
All headers are Architects Daughter.
All the rest of the text is Playfair Display.
The margin top and bottom for resources section headers were increased and decreased respectively to give more space.
The company logos were added to the resource cards with no maps.
An if/else statement was included so if a map attribute was present the view would create the iframe for the map otherwise it would look for a vetlogo. This prevented an empty iframe from appearing on the resource card.