Aligning vet logos to the right of the resource cards.
Why:
Having the logos aligned to one side looks more uniform. The right side of the card is emptier in most cases.

This commit:
Uses the float property to align the logo to the right of the card.
Padding has then been used to adjust the logo's position to look less abrupt.