Javascript fix and new partial for adoption application
Why:
The js for the adoption application form was only working when the page was refreshed but not when the route was followed from home to the form. The view of the adoption application is very information heavy.

This PR:
Updates the js so that the event triggers on every page load.
The bulk of the adoption application has been moved to a partial _form_fields.html.erb. This clears the view so it is less overwhelming and has more clarity on it's purpose.
Deleted some trailing whitespace and ordered some css properties.