Content updated for volunteer page
Content updated and organized.
Formatted using p tags
volunteers.scss file has been created and a few styles have been added to align the text in the center of the page and format the headers of each mini section.
Started to put infrastructure in place for linking to the foster application form.