Updated image names to follow a sequence.
Why:
I want file names to be informative to anyone who will look at the code. It makes sense to me for all rat images to be called rat followed by an index.

This commit:
Renames rat.jpg to rat-0.jpg
Renames gw.jpg to rat-1.jpg
The files and the pages where the pictures are called have been updated.