Links to forms corrected.
A link has been set up to take the user to the foster application form.
The adoption application form link has been corrected so that the link is within the sentence not after.
The css class name for the main div on the volunteer page has been modified to be volunteer-text-image.