Volunteer routes amended
Made volunteer a singular resource. This enables volunteer to have a show action without requiring an id.
Index was currently being used. This has been changed in the controller and view to show.
A few updates have been made to the h2 for the volunteer page. Top and bottom margins have been modified so the page looks more even.