Moved error messages partial
The adoptions application form has it's own partial for rendering error messages. I will need one for the foster application too.

Instead of duplicating the file I have modified the error partial to call the errors on an object rather than specifically the @adoption_application. The object is passed in when the partial is rendered. The partial has been moved to a separate folder in views called error_messages. This way the same partial can be rendered if needed by the foster application form.