Updated Foster Application to Volunteer Application
Why:
This application form covers the choice of fostering or being a quarantine home.
Need to make sure that applicants can commit to adoption appointments.

This commit:
Application has been renamed to Volunteer Application Form. This is a more descriptive title.
Renamed some data fields to be consistent with the Adoption Application Form.
Added a data columns/radio buttons/notes for adoption_appointments. This is to ensure that applicants are aware of the commitment of meeting with adopters for their rats.
Added an any questions and comments section so that an applicant can add extra information if desired.
Changed the work_location question to a radio button.