Hound corrections
Deleted test files as they are not in use.
In errors stylesheet: changed id selector to a class selector, replaced underscores with hyphens for field with errors selector. Changes were then reflected in the error_messages views.
In models, reduced line to under 80 characters.
Added flash.now to the controller as the flash messages were remaining after a page move.