Removed redundant fields made a couple of corrections
Ammended the flash messages. If using redirect there is no need to use flash.now. However if rendering a template flash.now is useful. Reflected this in the adoption application controller and the volunteer application folder.
Removed test/fixtures file, coffee file and helper file for the foster application as none of these are being used.
When rendering a file it is not necessary to put the full file extension. Just the file name and the folder path to it relative to the file which is rendering. The file extension has been deleted from the render call in adoption application view. Also switched to double quotes as that is my style pattern.