Validations and line length corrections
Pulled out the cell phone validation into it's own file. This allows it to be used for both models but only defined once.

Added a method in config/application that when field_with_errors is called it creates a new proc called field-with-errors. This allows me to ammend the css class to comply with my style guide.

Shortened several lines in adoption_application_controller, volunteer_application_controller and the volunteers/show view to be under 80 characters.