Preview HTML Mailer working for adoption application
The new mailer renders a partial for the form content which in turn renders a partial for each data entry.
This way the formating for table is defined only once in the data entry partial.
Originally the method for sending the email took in an argument called object. This has been renamed to be adoption_application. This is for clarity purposes.