Updated Radio Buttons for Adoption Application form
Why:
I noticed that when the data saved to the model there was no value saved for the questions asked using radio buttons. The emails being sent to the rescue therefore just had a blank cell for these questions.

This commit:
Changes the radio_button_tag to be f.radio_button. This links the radio button to the form and as a result the value chosen is saved with all the other inputs from the other form fields.

Side effects:
The way the arguments and options are written needed to be changed based on this.
Yes and no replaced with true and false.
Instead of writing just true, now have to include :checked => true in order to default select one option.
This caused the dependent fields to break as the data-radio-name now needs to include the form name.