Corrections and refactors
Improved the mailer views. The form fields have been pulled into a constant in the model. This constant is an array of strings. This can then be called by the mailer view and looped over. Each string can be used for the left column and then used to call the corresponding value from the adoption application. The html formatting can be called within the loop. This cuts out the use of two partials.

Updated my hash syntax to not use the =>
Removed some instances of .to_s as <%= %> does that anyway.
Renamed the mailer NewAdoptionApplicationMailer as it is handling a new adoption application and the email method is notification.
Shortened a few lines below the 80 character limit and replaced single quotes with double quotes.