New Volunteer Application Mailer Generated
Mailer has been generated.
It shares the admin_email method with the new adoption application mailer so this method has been moved to the application mailer file.
Created a constant FORM_FIELDS so that I can use it to create the html table in the mailer.
Created a preview for both html and text.
Updated the params in the controller as some options had been added to the migration but the controller had not been updated.