Added submission time of the form
Having a submission time on the form would help the rescue keep track of when they received it so they can follow up promptly.

Defined a method in app/helper that converts the created_at value to local time and then formats it to be easier for the reader.
Had to add_template_helper to the Mailer to access these methods.