Displaying monthly flyers
Ammended fliers to flyers. Decided that flyers is more common and have decided this is how I am going to refer to them. All files have been ammended to use this spelling.
Add 13 flyers to an assets file.
Created a FLYERS constant that converts all the flyers file names to strings.
The FLYERS constant can be called and iterated through to render the view.
N.B the flyers are being accessed but not loading correctly in the view.