Hooked up Donate button to Paypal
Why:
Currently Any Rat Rescue links to their paypal website to accept donations.

This commit:
Clicking the donate button will open a new browser tab which directs to paypal home.

Side effects:
As the text on the button is now a link the default styling for links applied. This meant the text was difficult to see.
Updated the styles for link which are direct descendants of buttons to have white text. On hover the text remains white but the button background color still changes.