Removed the dependency on the FLYERS constant
Why:
In order to add a new flyer to the website it would need to be added to the assets file but also the name would need to be put into the constant. It would need to be put in the correct order as well. It would be nice to only have to add the file to the assets folder.

This commit:
Includes the code to perform the following actions.
1 - The contents of the monthly_flyers folder within the assets folder are retrieved.
2 - Select only files with .pdf in the name
3 - Sort the files by date.
4 - Remove the .pdf file extension