Added thumbnails for the flyers
Why:
The page will look better with thumbnails of the flyers that can be clicked to enlarge.

This commit:
Adds png files of all the flyers to the assets folder. The png has the same name as the corresponding pdf. Therefore the same methods can be used to access the png as the pdf.
Formatted the text and thumbnails on the page.