Refactor of the newsletters controller
Why:
Both news and flyers were calling their own set of the same methods.

This commit:
The method chain is the same and the other than the first and final methods they do not need to know which folder they are dealing with. The first method searches through the directory of a folder which is provided by an instance variable defined in the final method. The final method is setting the array of either flyers or news and so can specify which folder the method chain should begin.