Added subscribe and email us link
Why:
Want it to be easy for visitors to subscribe to a mailing list. Also want visitors who want to post the flyer to contact us to let us know.

This commit:
Adds a subscribe button.
Adds an email us link.
Neither of these links are hooked up yet.
Centered the text as this page doesn't have blocks of text and justify wasn't looking quite right.