Created Newsletter Finder Model
Why:
The Newsletter controller was very logic heavy. It handled and had access to more information than it needed.

How:
Created a model NewsletterFinder.
Logic to find files, sort and format the names has been moved to the model.
The controller can call a new NewsletterFinder object and call the final method (either news or flyers) which will then proceed through the logic in the model.