Corrections
Deleted the newsletters/show.html.erb file. This file was added during the initial planning. It is not being used as each flyer opens as a pdf so does not need a view.
Removed an empty line from the migration file.
Changed the h1 tags to h2 tags. The text in these tags are sub headings not titles. This is flows better with the style of the overall document.