Added a validation for the message.
The message can't be blank.

On checking that the validation was working correctly I noticed that the error styling was not applying to it because the input is a textarea instead of textfield.

The errors.scss file has been updated to include textarea as a child of field-with-errors. It correctly displays if not message is written.