Removed the custom fonts
Why:
Bug was found that the text was not loading correctly on a redirect.

How:
Removed the custom fonts and reverted to the default fonts provided by bitters.
I have run through the sequence where the bug occured a couple of times and seen no occurance of the bug.

Side effects:
The resource cards with the logos instead of a map need tweaking as this new font is slightly larger and has pushed one of the logos slightly off the card. This will need to be addressed as it looks messy. Otherwise this change has not affected the layout of any other the other pages.